move into shrink and remove test cases
